Add test cases for "incorrect" configs
Test cases cover configs created with wrong/extra messages/types - proto uses the defined number to map fields, so as long as the request matches those it will process as expected.

Fix get file name so it doesn't return a path with no file name when the file name could not be created.

Test: run tests
Bug: 293957254
Change-Id: I119bfa5d91db38c14c73603634d27ff285120693