Improve test runtime by using stack samples and reducing timeouts
This change reduces the wait for the callback to 5 second increments
from 30 second increments, which is usually hit in the first or second
increment.

Also, this change reduces the amount of time we spend actually
collecting the profiles by replacing java heap dumps, which take longer
to collect, with one second long stack samples.

With these two changes, the longest-running tests in the module take
less than 11 secondsd to run - most take significantly less time.

Bug: 330089415
Bug: 330089393
Bug: 328948216
Bug: 328944189
Bug: 328929885
Bug: 328901441
Bug: 328907239
Bug: 328900747
Bug: 328900153
Bug: 328897712
Test: atest MtsProfilingModuleTests, locally verifying improved test
runtime
Change-Id: I2ba966d31798ccd41ed50e1547d23519ebc1af69

Change-Id: I01833819e32fdb9565db797df9531e04cbbf2777