Update perfetto permissions test to mock the Perfetto permission
It seems that Perfetto now has permissions to be called from the testing
app. Rather than write a test that enforces platform permission
structure (which is out of scope), rework this test to mock the lack of
Perfetto access so that we are testing that lack of Perfetto access is
handled correctly.

Bug: 329828599
Test: atest ProfilingServiceTests#testRequestProfiling_Allowed_PerfettoPermissions_Fails
Change-Id: I5e83979592a292e36504acec050168f8d3f05298