Setup for calling trace_redactor
Add the initial setup and flow for redacting traces before delivering to
app storage.  With this change when a trace is completed a new processes
will be created to run the trace_redactor binary.  The newly redacted
file will then be delivered to app storage and the unredacted file will
be deleted.

Bug: 327423523
Test: Atest CtsProfilingModuleTests
Change-Id: Ic8f11584b698650657329e0e2d8f32db63a3e14b