Adding preupload hooks to the Profiling module
This will allow Checkstyle and other hooks to run when we upload
changes, which should make code reviews cleaner.

Bug: 332784177
Test: Saw that checkstyle hooks ran on presubmit on this change if I
created a checkstyle error (moving an import). Verified that existing
presubmit hooks (like the one to have the Test: stanza in the commit
message) still ran despite not being specifically included in this
file.

Change-Id: Ieeb6f9ec6960c537ce73b5286ac7077ffeff66a6