Move obtaining package name to ProfilingManager
The current method doesn't handle all packages well, and getting a list of all packages associated with
a uid does not guarantee we're profiling/redacting for the right app. Obtain the package name from calling
application context instead and plumb through, making sure to verify on service side that the provided
package name is associated with the uid.

Bug: 345225912
Test: run ProfilingFrameworkTests and ProfilingServiceTests, check output results manually, confirm process name correct when manually logged
Flag: EXEMPT - bugfix
Change-Id: I06b48c076ea723d55710f463a09c437387817178