Add timestamp to system triggered session
Without the profiling start timestamp the record will be cleaned up any
time a cleanup is triggered, even if the session is currently being
processed.

This issue causes test flake as the cleanup timing can vary.

Also add a comment to the move to queue method to specify that
timestamp must be set.

Test: presubmit, abtd with cl passes testSystemTriggeredProfiling 100%,
      without cl that test has flake
Bug: 385338671
Bug: 383353460
Flag: EXEMPT - small bugfix
Change-Id: I802f1fa8ade6e91cdb9c0e98dd82b546d5d6df69