Modify debug retain results mode to retain all results
The previous version retained only the unredacted trace for debugging as
it is not directly delivered to the client. However, developers on user
builds are unable to easily access the redacted files from their own
apps storage. To make this easier, modify the command so that it stops
all results from being deleted, developers can then access the files as
they would most perfetto traces.

Test: run frameworks cts with config set to true, observe all results
      are retained. Run again with config false, observe results deleted.
Bug: 369436215
Flag: EXEMPT - debug command
Change-Id: I2cebc4cd8bb122bfc67f941328ae5cb0febe46ad