Fix potential case of multiple system triggered traces simultaneously
Currently, the system triggered trace can end up with more than one
instance if a debug triggered trace is in progress and the regular
system triggered runnable kicks in. This leaves instances untracked
and therefore not explicitly stoppable (thus ending after timeout
only). When running tests repeatedly, these leaked sessions results
in flake as further profiling cannot be started.

To fix this:
- Synchronize access to the system triggered profiling process
    object to ensure that antoher cannot be started while one is
    currently being started.
- Add an early return to the start system triggered method so that
    restarts won't be attempted while a profiling is in progress.

In addtion to this, change the log level for fail to start.

Test: https://android-build.corp.google.com/abtd/run/L40300030009396061/
Bug: 394352179
Flag: EXEMPT - small bugfix
Change-Id: I82c53ee9054186bf5d092009c8af8f20afea4e1e