Make temp files publicly readable if debug mode is enabled
Default permission is 600 and since owner is system, adb cannot
access the files. To remedy this, make the files publicly readable
if the debug mode is enabled. Access to the files is still limited
by sepolicy, so only system server, shell, and adbd will now be
able to access.

Bug: 369436215
Test: combined with sepolicy changes, pull files from dir.
Flag: EXEMPT - small bugfix
Change-Id: I252e2ba9844b1385f002ee9f3f37a99ecf09ad56