Fix exception handling of stopProfiling
ProfilingManager doesn't do anything with the exception since the
request comes from a CancellationSignal and we don't bother with a
dedicated API to report failure to stop the profiling. Furthermore,
calls from a handler thread don't expect this exception and therefore
are not currently handling it. Switch to handling failure gracefully
in all cases, if this happens the profiling will end naturally after
the time limit and will be delivered to the app, but won't be
capturing what they likely wanted.

Test: profiling mcts
Bug: 389096905
Flag: EXEMPT - bugfix
Change-Id: Ic21d81a0de15621d60001962f30fbcf8d59decdf