Refactor flagging
Currently, service cannot access non-exported flags. This is
problematic as functionality changes, which are mostly in service,
should use non-exported flags.

The issue arrises due to the use of an exported container which
tells the system not to generate accessors for non-exported flags.

Fix this by creating 2 containers, consistent with other modules,
an exported and not exported one. The naming scheme is reversed
(calling out not exported rather than not exported) as the current
(exported) dependency is already used outside the module.

Additionally, this change moves flags to their own directory for a
cleaner stucture. This is not required for the issue described above.

Bug: None
Test: add non-exported flag and access in both framework, service,
    and tests.
Flag: EXEMPT - refactor
Change-Id: Icc7dd5c06dde401217c4c43bc49de46ea3ab818d