Handle missing file before starting redaction.
Add a check in `handleRedactionRequiredResult` to ensure a filename
exists on the tracing session before attempting to create files for the
redaction process. If the filename is missing, set an error state and
advance the session.

This fail-fast check can reduce test flakiness when the following code
in `handleRedactionRequireResult` runs on a slower device.

Bug: 442827267
Test: atest android.profiling.cts.ProfilingServiceTests
Flag: EXEMPT BUGFIX
Change-Id: Iaf23906283058defb412059319e0708e74620982