Add e2e test for ANR trigger
This change introduces a new CTS test case to verify the functionality
of system-triggered profiling for ANR events. The test sets up an ANR
profiling trigger, intentionally causes an ANR in a test application by
blocking a broadcast receiver, and then verifies that a profiling result
is successfully generated and delivered.

Refactor using AmUtils to force stop a package.

Test: atest android.profiling.cts.ProfilingAppTriggeredTests
Bug: 377781177
Flag: TEST_ONLY

Change-Id: I5a60b2355271bd0d71c271644fa7dc7fb0e84cb6