Fix app crash tests flake
Add waitForCallback to testAddSpecificListenerNoCurrentListeners which
was exiting with a result still waiting, causing it to be returned in
the next test which was sometimes
testSystemTriggeredApplicationCrashNotProfilingEligible.

Add waits after setting the confg because sometimes the test was
progressing before the config update took effect. This caused flake as
the debug package may not be set yet causing rate limitng to fail.

Add overrides to java heap dump configs to ensure killswitch is always
off.

Adjust test wait times to use returned value and reduce padding time

Test: Run test suite 15x
Bug: 452371799
Flag: EXEMPT TEST_ONLY
Change-Id: Id398dc5eba4d68242671fadb3514da0eb2ce5b54