Add app cold start trigger
The trigger happens when the app start as a cold start. It starts a new
profiling session with 5 seconds duration to collect both system trace
and stack sampling. The perfetto trace buffer policy is set to `DISCARD`
so it retains trace since started and stops when the buffer is full.

Test: atest CtsProfilingModuleTests:android.profiling.cts.ProfilingFrameworkTests
Bug: 458051536
Flag: android.os.profiling.profiling_trigger_cold_start

Change-Id: Ibc4b56188450b4e0299eddf228bd44d4198e68a1