Fix CtsProfilingModuleTests AbstractMethodError
The `CtsProfilingModuleTests` were failing with an `AbstractMethodError`
related to `GeneratedMessageLite.dynamicMethod`. This was caused by a
conflict in jarjar rules, where the service module and the CTS test module were shading `com.google.protobuf` to different package names, leading to a runtime class incompatibility.

This change updates the `jarjar-rules.txt` for `CtsProfilingModuleTests` to align its protobuf jarjar rule with that of the service module.

Test: presubmit / abtd
Bug: 465804165
Fix: 468205293
Flag: EXEMPT PURE_REFACTOR

Change-Id: I11d478ad65f6a330faf3a452731fecd3f7cef4be